Macro-ls for symbolic input to ls
Now need a possibility to unquote and eval the input to ls.
A good solution would be to eval every list of ls' input.
Thus in:

	ls a (foo bar) c

a and c are not eval-ed (not searched for values among bound variables),
but (foo bar) is eval-ed to some result foobar symbol (or string?) and all three are passed to ls-func as:

	ls-func, quote (a foobar c)

	modified:   README.md
	modified:   ls.syntax