Remade , and : according to README, added \ and example0
Fixed tabs, added \ for no-call tabs.
Changed coma to nest symbols before the coma.
The colon nests after the colon until the end of the line.
Now need proper substitution for the case of:

	quote :
		a b
		c d
	=
	(quote ((a b) (c d)))

(Or make the colon multi-line?)

Noticed -- the parentheses are broken! (examples 4, 5, 7..)

The whole parsing should be redone...

	new file:   example0
	modified:   example1
	modified:   example4
	modified:   example5
	modified:   example7
	modified:   example8
	modified:   example9
	modified:   ls.syntax
	modified:   parsing_scripts.py